A simple rune comparison is implemented
Runes drawn will not automatically be resized to a predetermined
constant "runeSize" (found in Rune.java), currently set to 50. They will
also be moved to the center of this size (so the longest side of a rune
will be stretched to 50, the other side moved).

Runes can be loaded from Images that have complete black and white
colors and the same will apply here.

Runes can be compared using the "similarTo" method of Rune (maybe a
better name can be thought of). This currently uses an "average
distance" algorithm using a lookup table. Currenty the lookup table is
calculated every time the function is called (lol!) which is ofc super
ineffective, but since the size is only 50 it's not TOO bad.

Boom shakalacka.